backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
sqlalchemy
sqlite-utils
python-dotenv
